+++ Healthy/Domain/Sources/Water/Model/WaterModel.swift
@@ -10,9 +10,9 @@ import Foundation
 public struct WaterModel {
     
     public var unit: String
-    public var goal: Int
+    public var goal: Float
     
-    public init(unit: String, goal: Int) {
+    public init(unit: String, goal: Float) {
         self.unit = unit
         self.goal = goal
     }
